% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dividir.R
\name{divisao}
\alias{divisao}
\title{Função para dividir}
\usage{
divisao(a, b)
}
\description{
Essa função serve para dividir dois numeros
}
\details{
@param a Um numero
  @param b um outro numero

  @examples
  somar(2,3)

  @export
}
