% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Fibonacci.R
\name{fibonacci}
\alias{fibonacci}
\title{Sequencia de Fibonacci}
\usage{
fibonacci(x, arg = T)
}
\arguments{
\item{x}{e um numero natural.}

\item{arg}{e o tipo de saida.}
}
\value{
se arg=T, retornara o elemento de fibonacci na posicao x.

se arg=F, retornara um vetor com x primeiros elementos de Fibonacci.
}
\description{
Essa funcao retorna um vetor com x primeiros elementos de fibonacci 
ou o elemento na posicao x.
}
