% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multiplicar.R
\name{multiplicaçao}
\alias{multiplicaçao}
\title{Função para multiplicar}
\usage{
multiplica攼㸷ao(a, b)
}
\description{
Essa função serve para multiplicar dois numeros
}
\details{
@param a Um numero
  @param b um outro numero

  @examples
  somar(2,3)

  @export
}
