% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Somar.R
\name{somar}
\alias{somar}
\title{Função para somar}
\usage{
somar(a, b)
}
\description{
Essa função serve para somar dois numeros
}
\details{
@param a Um numero
  @param b um outro numero

  @examples
  somar(2,3)

  @export
}
