% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Subtrair.R
\name{subtraçao}
\alias{subtraçao}
\title{Função para subtrair}
\usage{
subtra攼㸷ao(a, b)
}
\description{
Essa função serve para subtrair dois numeros
}
\details{
@param a Um numero
  @param b um outro numero

  @examples
  somar(2,3)

  @export
}
