\name{Teste-package}
\alias{Teste-package}
\alias{Teste}
\docType{package}
\title{
\packageTitle{Teste}
}
\description{
\packageDescription{Teste}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{Teste}
\packageIndices{Teste}
~~ An overview of how to use the package, including the most important functions ~~
}
\author{
\packageAuthor{Teste}

Maintainer: \packageMaintainer{Teste}
}
\references{
~~ Literature or other references for background information ~~
}
~~ Optionally other standard keywords, one per line, from file KEYWORDS in the R ~~
~~ documentation directory ~~
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
~~ simple examples of the most important functions ~~
}
